feat: :sparkles: Adding and modifying files - Added new folder Prize with its own files - Added folder controller and model to folder Prize - Added init files to controller and model folders - Modifying files to cosplay folder

+++ src/main/java/com/u2team/huansync/activity/cosplay/categorycosplay/controller/CategoryCosplayDAO.java
@@ -3,7 +3,6 @@ package com.u2team.huansync.activity.cosplay.categorycosplay.controller;
 import com.u2team.huansync.activity.cosplay.categorycosplay.model.CategoryCosplay;
 import com.u2team.huansync.persistence.BDConnection;
 
-
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
@@ -15,14 +14,13 @@ public class CategoryCosplayDAO {
 
     public List<CategoryCosplay> getAllCategories() {
         List<CategoryCosplay> categories = new ArrayList<>();
-
         try (Connection connection = BDConnection.MySQLConnection()) {
             String sql = "SELECT * FROM tbl_categoryCosplay";
             try (Statement statement = connection.createStatement();
                  ResultSet resultSet = statement.executeQuery(sql)) {
                 while (resultSet.next()) {
                     CategoryCosplay category = new CategoryCosplay();
-                    category.setCategoryCosplayId(resultSet.getInt("categoryCosplayId"));
+                    category.setCategoryCosplayId(resultSet.getLong("categoryCosplayId"));
                     category.setNameCategoryCosplay(resultSet.getString("nameCosplay"));
                     categories.add(category);
                 }
@@ -30,7 +28,6 @@ public class CategoryCosplayDAO {
         } catch (SQLException e) {
             e.printStackTrace();
         }
-
         return categories;
     }
 
@@ -40,10 +37,9 @@ public class CategoryCosplayDAO {
             try (PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
                 statement.setString(1, category.getNameCategoryCosplay());
                 statement.executeUpdate();
-
                 try (ResultSet generatedKeys = statement.getGeneratedKeys()) {
                     if (generatedKeys.next()) {
-                        category.setCategoryCosplayId(generatedKeys.getInt(1));
+                        category.setCategoryCosplayId(generatedKeys.getLong(1));
                     }
                 }
             }
@@ -57,7 +53,7 @@ public class CategoryCosplayDAO {
             String sql = "UPDATE tbl_categoryCosplay SET nameCosplay = ? WHERE categoryCosplayId = ?";
             try (PreparedStatement statement = connection.prepareStatement(sql)) {
                 statement.setString(1, category.getNameCategoryCosplay());
-                statement.setInt(2, category.getCategoryCosplayId());
+                statement.setLong(2, category.getCategoryCosplayId());
                 statement.executeUpdate();
             }
         } catch (SQLException e) {
@@ -65,11 +61,11 @@ public class CategoryCosplayDAO {
         }
     }
 
-    public void deleteCategory(int categoryId) {
+    public void deleteCategory(long categoryId) {
         try (Connection connection = BDConnection.MySQLConnection()) {
             String sql = "DELETE FROM tbl_categoryCosplay WHERE categoryCosplayId = ?";
             try (PreparedStatement statement = connection.prepareStatement(sql)) {
-                statement.setInt(1, categoryId);
+                statement.setLong(1, categoryId);
                 statement.executeUpdate();
             }
         } catch (SQLException e) {
@@ -77,4 +73,3 @@ public class CategoryCosplayDAO {
         }
     }
 }
-
